requirements.txt
# Core Scraping Dependencies
scrapy==2.11.0
selenium==4.15.2
webdriver-manager==4.0.1

# AI and NLP Dependencies
langchain==0.0.350
langchain-openai==0.0.2
openai==1.3.7
openrouter==0.1.0

# Database Dependencies
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Dashboard and UI
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.3
numpy==1.25.2

# Data Processing
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# Scheduling and Utilities
apscheduler==3.10.4
python-dotenv==1.0.0
pydantic==2.5.0

# Development and Testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0 